[modified] Move all landing and home composable(s) to HomeActivity.kt, so that MainActivity.kt is now only for navigation.
