+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">funHacks2022</string>
+    <string name="title_activity_home">HomeActivity</string>
+    <string name="title_activity_driving">DrivingActivity</string>>
 </resources>
